Create a debug spec handler with C entry points
Introduce debug specifications with C entry points to facilitate control
of the C++ objects. This will allow the runtime control of whether a
message can log via debug specifications (debug specs) passed in and
parsed using the notation :
[<FILE>::][<TAG>[::<TAG>]::]#[, ...]

Where the following are as defined as :
  ::      delimiter between tokens
  ,       delimiter between specifications
  <FILE>  token: filename to restrict spec to
  <TAG>	  token: custom tag to restrict spec to
  #       token: debug level, always required on a spec
  ...     additional specifications following same syntax
  [ ]     optional token or input

Using these debug specs a message can be tested for validity using the
single function call `canLog()` or `DebugSpecHandler_canLog()` when
providing its respective tags, filename, and debug level.

The evaluation logic of whether a particular message can log is :
  [<FILE>::][<TAG>[::<TAG>]::]#, [<FILE>::][<TAG>[::<TAG>]::]#, ...
      ^---and--^--and--^--and-^      ^---and--^--and--^--and-^  ...
           ^-----------or----------------^                      ...

Using this logic we have the following constraints:
* All restrictions / tags / level within a single specification must be
  satisfied for a message to be valid for this particular specification
  (and logic)
* Only one specification needs to be valid from the set of all debug
  specifications to be valid for output (or logic)