Reverse the order of log levels to allow increasing levels of debug
By switching the log level direction, priority is given to the lowest
value. To support increased levels of debug verbosity, the open
positive end is set to debug level and critical is set to the minimum
value (maximum priority) of 1. This allows for one level greater than
critical if needed without redefining these values.