Create, manage, and use the debug spec handler
Create a debug spec handler specific to the core within the logging
module and add ensure backward compatibility with existing calls by
setting the default level to MPAS_LOG_OUT.

Add optional arguments to the mpas_log_write() call, allowing for the
provision of a filename and any associated tags of this particular write.
Within the function, call DebugSpecHandler_canLog to check if this write
is valid. There is no affect on existing write calls due to the default
setting of MPAS_LOG_OUT.

On clean up, ensure the destructor is called for the underlying C++
object.