Add a logging extension module to initialize from namelist options
Due to the nature of Fortran module compilation, the logging module
cannot call any MPAS pool routines to fetch namelist inputs. To
workaround this issue and allow runtime control of logging debug specs,
a separate module will handle interfacing the debug spec handler of a
particular domain to the appropriate namelist option. This breaks the
circular dependency so long as mpas_log never depends on the extension
module.