Remove 'pointer' attribute from 'inlist' argument of mpas_dmpar_scatter_ints
The 'pointer' attribute is unnecessary for the 'inlist' argument, and requiring
the inlist argument to be a pointer precluded the use of mpas_dmpar_scatter_ints
for arrays that are not pointers, e.g., allocatable arrays or array
constructors. Additionally, since the inlist argument is used only as the first
argument to MPI_Scatterv (i.e., as the 'sendbuf' argument of MPI_Scatterv), it
can be declared as an intent(in) in the mpas_dmpar_scatter_ints routine.

The changes in this commit have also been found to resolve runtime errors
occurring with certain compiler and MPI library combinations, specifically,
nvfortran and OpenMPI 5.x.